feat: update discord.js to 14.19.x
- add support to components v2
