+++ package.json
@@ -40,14 +40,14 @@
     "tslib": "^2.8.1"
   },
   "devDependencies": {
-    "@types/node": "^22.13.10",
+    "@types/node": "^22.15.3",
     "@types/source-map-support": "^0.5.10",
-    "discord.js": "^14.18.0",
-    "eslint": "^9.22.0",
+    "discord.js": "^14.19.1",
+    "eslint": "^9.25.1",
     "gen-esm-wrapper": "^1.1.3",
     "globals": "^16.0.0",
-    "typescript": "^5.8.2",
-    "typescript-eslint": "^8.27.0"
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.31.1"
   },
   "peerDependencies": {
     "discord.js": "^14.x"
